requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Data Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Task Queue
redis==5.0.1
arq==0.25.0

# LLM Integration
openai==1.10.0
instructor==0.5.2
rapidfuzz==3.6.1
pgvector==0.2.5
alembic==1.13.1
anthropic==0.39.0
google-generativeai==0.8.3
prometheus-client==0.20.0
structlog==24.4.0
httpx==0.27.2

# Development & Testing
pytest==8.0.0
pytest-asyncio==0.23.3
ruff==0.1.14
coverage==7.4.1
